changing link script to remove files after it knows the command is valid.

+++ link.sh
@@ -1,23 +1,31 @@
 #!/bin/bash
+
+if ! [ -z $1 ]
+then
+	if [ -f ~/dotfiles/$1_bashrc ]
+	then
+		echo "Linking for $1 machine."
+		rm -f ~/.bashrc
+		ln -s ~/dotfiles/$1_bashrc ~/.bashrc
+	else
+		echo "That bashrc doesn't exist..."
+		exit 1
+	fi
+else
+	echo "Linking for personal machine."
+	rm -f ~/.bashrc
+	ln -s ~/dotfiles/bashrc ~/.bashrc
+fi
+
 pushd ~/ > /dev/null
 
 # remove old files or links
-rm -f .bashrc
 rm -f .bash_profile
 rm -f .gitconfig
 rm -f .vimrc
 rm -f .tmux.conf
 
 popd > /dev/null
-
-if ! [ -z $1 ]
-then
-	echo "Linking for $1 machine."
-	ln -s ~/dotfiles/$1_bashrc ~/.bashrc
-else
-	echo "Linking for personal machine."
-	ln -s ~/dotfiles/bashrc ~/.bashrc
-fi
 
 ln -s ~/.bashrc ~/.bash_profile
 ln -s ~/dotfiles/vimrc ~/.vimrc
